Adding movement and rotation to simple-three example

+++ examples/simple-three/src/lib/components/model3.data.ts
@@ -1,9 +1,9 @@
 import { component } from '@dark-star/ecs';
-import { int32, string16 } from '@dark-star/shared-object';
+import { float64, string16 } from '@dark-star/shared-object';
 
 @component()
 export class Model3 {
-	@int32()
+	@float64()
 	scale: number = 1;
 
 	@string16(30)
